updated imports, and restructured
