--- web-app/src/app.py
@@ -1,95 +1,0 @@
-from flask import Flask, render_template, request, redirect, url_for, flash
-from flask_login import flask_login, current_user, login_required
-from pymongo import MongoClient, pymongo
-import os
-import certifi
-
-app = Flask(__name__)
-
-#login 
-login_manager = flask_login.LoginManager()
-login_manager.init_app(app)
-login_manager.login_view = "login"
-
-#mongodb client
-client = pymongo.MongoClient(os.getenv("MONGO_URI"), tlsCAFile=certifi.where())
-
-@app.route("/home")
-@login_required
-def index():
-    """Render the homepage."""
-    return render_template("index.html")
-
-
-@app.route("/", methods=['GET', 'POST'])
-def login():
-    if request.method == 'POST':
-        username = request.form.get('username')
-        entered_pw = request.form.get('password')
-
-        if not username or not entered_pw:
-            flash("Username and password are required", "danger")
-            return redirect(url_for("login"))
-    
-        user_doc = users.find_one({"username": username})
-        #password = users.find({"username": username})
-
-        if user_doc and (entered_pw == user_doc["password"]): #correct credentials
-            user = User(user_doc)
-            #user.username = username
-            flask_login.login_user(user)
-            session["user"] = user.id
-            return redirect(url_for("home"))
-        else:
-            flash("Invalid username or password", "danger")
-            return render_template("login.html")
-        
-    return render_template("login.html")
-
-class User(flask_login.UserMixin):
-    def __init__(self, user_doc):
-        self.id = str(user_doc["_id"])
-        self.username = user_doc["username"]
-        self.password = user_doc["password"]
-
-    @staticmethod
-    def is_authenticated():
-        return True
-
-    @staticmethod
-    def is_active():
-        return True
-
-    @staticmethod
-    def is_anonymous():
-        return False
-
-    def get_id(self):
-        return self.id
-
-
-@login_manager.user_loader
-def user_loader(user_id):
-    user_doc = users.find_one({"_id": ObjectId(user_id)})
-    if not user_doc:
-        return
-    
-    return User(user_doc)
-
-
-@login_manager.request_loader
-def request_loader(request):
-    username = request.form.get('username')
-
-    if not username:
-        return
-    
-    user_doc = users.find_one({"username": username})
-
-    if not user_doc: #user doesn't exist
-        return None
-
-    return User(user_doc)
-
-if __name__ == "__main__":
-    app.run(debug=True)
