+++ web-app/templates/tracker.html
@@ -1,0 +1,18 @@
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="styles.css">
+    <title>Tracker</title>
+</head>
+
+
+<body>
+    <div class="container">
+        <h1>Track Your Emotions</h1>
+
+        <br/>
+    </div>
+</body>
+</html>
